♻️  case detail refactoring
