+++ src/utilities/camelCaseToNormal.js
@@ -1,0 +1,3 @@
+export default function (string) {
+  return string.replace(/([A-Z])/g, ' $1').replace(/^./, function(str){ return str.toUpperCase(); })
+}
