+++ src/utilities/index.js
@@ -1,7 +1,9 @@
 import dateMapper from "./dateMapper";
 import debounce from "./debounce.js";
+import camelCaseToNormal from "./camelCaseToNormal.js";
 
 export default {
   dateMapper,
   debounce,
+  camelCaseToNormal,
 }
